feat: add History tab to SourceDetailView via shared DetailInspectorView + ProvenancePanel
- Extract DetailInspectorView, InspectorTab, ProvenancePanel from
  PageDetailView into Sources/WikiFS/Detail/ (shared components)
- Create ProvenanceEntry display model; both PageOrigin and SourceOrigin
  project to it via .provenanceEntry
- Extend SourceOrigin with versionID, agentKind, runTitle fields
- Add sourceEditHistory(sourceID:) to WikiStore protocol, GRDBWikiStore,
  and WikiStoreModel wrapper (read-only, mirrors pageEditHistory)
- Update sourceOrigin query + originFrom(row:) decoder to select
  sv.id, a.kind, and the chat-title runTitle subquery
- Wire up SourceDetailView with the tabbed inspector (Outline + History),
  using source-specific @AppStorage keys to avoid desync with page's keys
- PageDetailView updated to use the shared DetailInspectorView with
  @ViewBuilder outline closure

Closes the gap where SourceDetailView had an outline pane but no History
tab — sources now show the same provenance UI (date-first, operation badges,
clickable rows) as pages.
